requirements.txt
streamlit>=1.36.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
python-docx>=1.1.2